Add files via upload
Embark on a journey into the depths of visual intelligence with DeepVision Navigator, a program leveraging YOLOv3 for object detection in images. Witness the magic of computer vision as it identifies and labels objects with precision, bringing clarity to visual landscapes. Experience the fusion of technology and perception, exploring the unseen through code.