Add files via upload
Embark on a journey through the labyrinth of Amazon customer sentiment with our project, where millions of reviews converge to unveil hidden insights. Harnessing the power of fastText, we decode the nuanced emotions behind each star rating, offering a glimpse into the collective voice of the marketplace. Join us as we navigate the terrain of consumer sentiment, revealing the stories behind the stars.